Add initial integration for the new cody web
