+++ client/web/src/cody/chat/new-chat/components/chat-ui/ChatUi.tsx
@@ -1,0 +1,11 @@
+import type { FC } from 'react'
+
+import classnames from 'classnames'
+
+import { CodyWebChat } from '@sourcegraph/cody-web'
+
+import styles from './ChatUI.module.scss'
+
+export const ChatUi: FC<{ className?: string }> = props => {
+    return <CodyWebChat className={classnames(styles.chat, props.className)} />
+}
